more routes moved, limited testing
