+++ hashview.rb
@@ -20,6 +20,6 @@ class HashView < Sinatra::Application
   end
 end
 
-#require_relative 'models/init'
+require_relative 'models/master'
 require_relative 'helpers/init'
 require_relative 'routes/init'
