+++ models/master.rb
