+++ routes/init.rb
@@ -5,4 +5,8 @@ require_relative 'tasks'
 require_relative 'jobs'
 require_relative 'settings'
 require_relative 'wordlists'
+require_relative 'hashfiles'
+require_relative 'analytics'
+require_relative 'search'
+require_relative 'downloads'
 #require_relative 'main'
