+++ routes/tasks.rb
@@ -6,28 +6,28 @@ class HashView < Sinatra::Application
   
     haml :task_list
   end
-
+  
   get '/tasks/delete/:id' do
     varWash(params)
-
+  
     @job_tasks = Jobtasks.all(task_id: params[:id])
     unless @job_tasks.empty?
       flash[:error] = 'That task is currently used in a job.'
       redirect to('/tasks/list')
     end
-
+  
     @task = Tasks.first(id: params[:id])
     @task.destroy if @task
-
+  
     redirect to('/tasks/list')
   end
-
+  
   get '/tasks/edit/:id' do
     varWash(params)
     @task = Tasks.first(id: params[:id])
     @wordlists = Wordlists.all
     @settings = Settings.first
-
+  
     if @task.hc_attackmode == 'combinator'
       @combinator_wordlists = @task.wl_id.split(',')
       if @task.hc_rule =~ /--rule-left=(.*) --rule-right=(.*)/
@@ -39,32 +39,32 @@ class HashView < Sinatra::Application
         @combinator_right_rule = $1
       end
     end
-
+  
     @rules = []
     # list wordlists that can be used
     Dir.foreach('control/rules/') do |item|
       next if item == '.' || item == '..'
         @rules << item
     end
-
+  
     haml :task_edit
   end
-
+  
   post '/tasks/edit/:id' do
     varWash(params)
     if !params[:name] || params[:name].nil?
       flash[:error] = 'The task requires a name.'
       redirect to("/tasks/edit/#{params[:id]}")
     end
-
+  
     settings = Settings.first
     wordlist = Wordlists.first(id: params[:wordlist])
-
+  
     if settings && !settings.hcbinpath
       flash[:error] = 'No hashcat binary path is defined in global settings.'
       redirect to('/settings')
     end
-
+  
     # must have two word lists
     if params[:attackmode] == 'combinator'
       wordlist_count = 0
@@ -83,12 +83,12 @@ class HashView < Sinatra::Application
           end
         end
       end
-
+  
       if wordlist_count != 2
         flash[:error] = 'You must specify at exactly 2 wordlists.'
         redirect to("/tasks/edit/#{params[:id]}")
       end
-
+  
       if params[:combinator_left_rule] && !params[:combinator_left_rule].empty? && params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
         rule_list = '--rule-left=' + params[:combinator_left_rule] + ' --rule-right=' + params[:combinator_right_rule]
       elsif params[:combinator_left_rule] && !params[:combinator_left_rule].empty?
@@ -104,7 +104,7 @@ class HashView < Sinatra::Application
     task.name = params[:name]
   
     task.hc_attackmode = params[:attackmode]
-
+  
     if params[:attackmode] == 'dictionary'
       task.wl_id = wordlist.id
       task.hc_rule = params[:rule]
@@ -114,33 +114,34 @@ class HashView < Sinatra::Application
       task.hc_rule = 'NULL'
       task.hc_mask = params[:mask]
     elsif params[:attackmode] == 'combinator'
-      task.wl_id = wordlist_list
+      task.wl_id = wordlist_list 
       task.hc_rule = rule_list
       task.hc_mask = 'NULL'
     end
     task.save
-
+  
     redirect to('/tasks/list')
-
+  end
+  
   get '/tasks/create' do
     varWash(params)
     @settings = Settings.first
-
+  
     # TODO present better error msg
     flash[:warning] = 'You must define hashcat\'s binary path in global settings first.' if @settings && @settings.hcbinpath.nil?
-
+  
     @rules = []
     # list wordlists that can be used
     Dir.foreach('control/rules/') do |item|
       next if item == '.' || item == '..'
         @rules << item
     end
-
+  
     @wordlists = Wordlists.all
-
+  
     haml :task_edit
   end
-
+  
   post '/tasks/create' do
     varWash(params)
     settings = Settings.first
@@ -153,7 +154,7 @@ class HashView < Sinatra::Application
       flash[:error] = 'You must provide a name for your task!'
       redirect to('/tasks/create')
     end
-
+  
     @tasks = Tasks.all(name: params[:name])
     unless @tasks.nil?
       @tasks.each do |task|
@@ -163,9 +164,9 @@ class HashView < Sinatra::Application
         end
       end
     end
-
+  
     wordlist = Wordlists.first(id: params[:wordlist])
-
+  
     # mask field cannot be empty
     if params[:attackmode] == 'maskmode'
       if !params[:mask] || params[:mask].empty?
@@ -173,7 +174,7 @@ class HashView < Sinatra::Application
         redirect to('/tasks/create')
       end
     end
-
+  
     # must have two word lists
     if params[:attackmode] == 'combinator'
       wordlist_count = 0
@@ -192,12 +193,12 @@ class HashView < Sinatra::Application
           end
         end
       end
-
+  
       if wordlist_count != 2
         flash[:error] = 'You must specify at exactly 2 wordlists.'
         redirect to('/tasks/create')
       end
-
+  
       if params[:combinator_left_rule] && !params[:combinator_left_rule].empty? && params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
         rule_list = '--rule-left=' + params[:combinator_left_rule] + ' --rule-right=' + params[:combinator_right_rule]
       elsif params[:combinator_left_rule] && !params[:combinator_left_rule].empty?
@@ -208,12 +209,12 @@ class HashView < Sinatra::Application
         rule_list = ''
       end
     end
-
+  
     task = Tasks.new
     task.name = params[:name]
-
+  
     task.hc_attackmode = params[:attackmode]
-
+  
     if params[:attackmode] == 'dictionary'
       task.wl_id = wordlist.id
       task.hc_rule = params[:rule]
@@ -224,9 +225,9 @@ class HashView < Sinatra::Application
       task.hc_rule = rule_list
     end
     task.save
-
+  
     flash[:success] = "Task #{task.name} successfully created."
-
+  
     redirect to('/tasks/list')
   end
 end
